Move fetch, add, and remove favorite movies to one redux lifecycle, added button for removing a favorite movie (without reloading), and redirect to login when trying to view profile page when logged-out.
