+++ src/reducers/index.js
@@ -1,5 +1,4 @@
 export { default as allMovies } from "../store/utilities/allMovies";
 export { default as user } from "../store/utilities/user";
 export { default as movie } from "../store/utilities/movie";
-export { default as favorite } from "../store/utilities/favorite";
-export { default as favoriteMovies } from "../store/utilities/favoriteMovies";
+export { default as favoriteMovies } from "../store/utilities/allFavoriteMovies";
