+++ src/thunks/index.js
@@ -1,5 +1,4 @@
 export * from "../store/utilities/allMovies";
 export * from "../store/utilities/user";
 export * from "../store/utilities/movie";
-export * from "../store/utilities/favorite";
-export * from "../store/utilities/favoriteMovies";
+export * from "../store/utilities/allFavoriteMovies";
